+++ src/test/java/testmod/scenario/package-info.java
@@ -1,0 +1,4 @@
+@NonnullDefault
+package testmod.scenario;
+
+import org.lwjgl.system.NonnullDefault;